refactor(tools): split admin meta builders
